feat: Подключил моки к `index.tsx`, создал компоненты `reviews` и `map` refactor: Обучил `offer-page` применять данные из пропсов
