+++ src/components/map/map.tsx
@@ -1,0 +1,7 @@
+function Map(): JSX.Element {
+  return (
+    <section className="offer__map map"></section>
+  );
+}
+
+export default Map;
