+++ src/components/reviews/index.ts
@@ -1,0 +1,2 @@
+import Reviews from './reviews';
+export default Reviews;
