+++ src/index.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './components/app';
 import { Settings } from './const';
+import mockOffers from './mocks/offers';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
@@ -9,6 +10,6 @@ const root = ReactDOM.createRoot(
 
 root.render(
   <React.StrictMode>
-    <App placesCount={Settings.PlacesCount}/>
+    <App placesCount={Settings.PlacesCount} offers={mockOffers}/>
   </React.StrictMode>
 );
