+++ src/pages/offer-page/offer-page.tsx
@@ -1,9 +1,38 @@
 import { Helmet } from 'react-helmet-async';
-import { getAuthorizationStatus } from '../../authorization-status';
-import { AuthorizationStatus } from '../../const';
+import { Offer, Offers } from '../../types/offer';
+import { useParams } from 'react-router-dom';
+import NotFoundPage from '../not-found-page';
+import Reviews from '../../components/reviews';
+import Map from '../../components/map';
 
-function OfferPage() {
-  const authorizationStatus = getAuthorizationStatus();
+type OfferPageProps = {
+  offers: Offers;
+};
+
+function OfferPage({ offers }: OfferPageProps) {
+  const { id } = useParams();
+  const currentOffer: Offer | undefined = offers.find(
+    (offer: Offer) => offer.id === id
+  );
+
+  if (!currentOffer) {
+    return <NotFoundPage />;
+  }
+
+  const {
+    title,
+    images,
+    isPremium,
+    rating,
+    type,
+    maxAdults,
+    bedrooms,
+    price,
+    goods,
+    host,
+    description,
+  } = currentOffer;
+  const ratingWidthStyle = `${rating * (100 / 5)}%`;
 
   return (
     <>
@@ -12,61 +41,32 @@ function OfferPage() {
       </Helmet>
       <main className="page__main page__main--offer">
         <section className="offer">
-          <div className="offer__gallery-container container">
-            <div className="offer__gallery">
-              <div className="offer__image-wrapper">
-                <img
-                  className="offer__image"
-                  src="img/room.jpg"
-                  alt="Photo studio"
-                />
-              </div>
-              <div className="offer__image-wrapper">
-                <img
-                  className="offer__image"
-                  src="img/apartment-01.jpg"
-                  alt="Photo studio"
-                />
-              </div>
-              <div className="offer__image-wrapper">
-                <img
-                  className="offer__image"
-                  src="img/apartment-02.jpg"
-                  alt="Photo studio"
-                />
-              </div>
-              <div className="offer__image-wrapper">
-                <img
-                  className="offer__image"
-                  src="img/apartment-03.jpg"
-                  alt="Photo studio"
-                />
-              </div>
-              <div className="offer__image-wrapper">
-                <img
-                  className="offer__image"
-                  src="img/studio-01.jpg"
-                  alt="Photo studio"
-                />
-              </div>
-              <div className="offer__image-wrapper">
-                <img
-                  className="offer__image"
-                  src="img/apartment-01.jpg"
-                  alt="Photo studio"
-                />
+          {images.length > 0 ? (
+            <div className="offer__gallery-container container">
+              <div className="offer__gallery">
+                {images.map((src, index) => (
+                  // eslint-disable-next-line react/no-array-index-key
+                  <div className="offer__image-wrapper" key={index}>
+                    <img
+                      className="offer__image"
+                      src={src}
+                      alt="Photo studio"
+                    />
+                  </div>
+                ))}
               </div>
             </div>
-          </div>
+          ) : null}
+
           <div className="offer__container container">
             <div className="offer__wrapper">
-              <div className="offer__mark">
-                <span>Premium</span>
-              </div>
+              {isPremium ? (
+                <div className="offer__mark">
+                  <span>Premium</span>
+                </div>
+              ) : null}
               <div className="offer__name-wrapper">
-                <h1 className="offer__name">
-                  Beautiful &amp; luxurious studio at great location
-                </h1>
+                <h1 className="offer__name">{title}</h1>
                 <button className="offer__bookmark-button button" type="button">
                   <svg className="offer__bookmark-icon" width="31" height="33">
                     <use xlinkHref="#icon-bookmark"></use>
@@ -76,39 +76,37 @@ function OfferPage() {
               </div>
               <div className="offer__rating rating">
                 <div className="offer__stars rating__stars">
-                  <span style={{ width: '80%' }}></span>
+                  <span style={{ width: ratingWidthStyle }}></span>
                   <span className="visually-hidden">Rating</span>
                 </div>
-                <span className="offer__rating-value rating__value">4.8</span>
+                <span className="offer__rating-value rating__value">
+                  {rating}
+                </span>
               </div>
               <ul className="offer__features">
                 <li className="offer__feature offer__feature--entire">
-                  Apartment
+                  {type}
                 </li>
                 <li className="offer__feature offer__feature--bedrooms">
-                  3 Bedrooms
+                  {bedrooms} Bedrooms
                 </li>
                 <li className="offer__feature offer__feature--adults">
-                  Max 4 adults
+                  Max {maxAdults} adults
                 </li>
               </ul>
               <div className="offer__price">
-                <b className="offer__price-value">&euro;120</b>
+                <b className="offer__price-value">&euro;{price}</b>
                 <span className="offer__price-text">&nbsp;night</span>
               </div>
               <div className="offer__inside">
                 <h2 className="offer__inside-title">What&apos;s inside</h2>
                 <ul className="offer__inside-list">
-                  <li className="offer__inside-item">Wi-Fi</li>
-                  <li className="offer__inside-item">Washing machine</li>
-                  <li className="offer__inside-item">Towels</li>
-                  <li className="offer__inside-item">Heating</li>
-                  <li className="offer__inside-item">Coffee machine</li>
-                  <li className="offer__inside-item">Baby seat</li>
-                  <li className="offer__inside-item">Kitchen</li>
-                  <li className="offer__inside-item">Dishwasher</li>
-                  <li className="offer__inside-item">Cabel TV</li>
-                  <li className="offer__inside-item">Fridge</li>
+                  {goods.map((god, index) => (
+                    // eslint-disable-next-line react/no-array-index-key
+                    <li className="offer__inside-item" key={index}>
+                      {god}
+                    </li>
+                  ))}
                 </ul>
               </div>
               <div className="offer__host">
@@ -117,206 +115,26 @@ function OfferPage() {
                   <div className="offer__avatar-wrapper offer__avatar-wrapper--pro user__avatar-wrapper">
                     <img
                       className="offer__avatar user__avatar"
-                      src="img/avatar-angelina.jpg"
+                      src={host.avatarUrl}
                       width="74"
                       height="74"
                       alt="Host avatar"
                     />
                   </div>
-                  <span className="offer__user-name">Angelina</span>
-                  <span className="offer__user-status">Pro</span>
+                  <span className="offer__user-name">{host.name}</span>
+                  {host.isPro ? (
+                    <span className="offer__user-status">Pro</span>
+                  ) : null}
                 </div>
                 <div className="offer__description">
-                  <p className="offer__text">
-                    A quiet cozy and picturesque that hides behind a a river by
-                    the unique lightness of Amsterdam. The building is green and
-                    from 18th century.
-                  </p>
-                  <p className="offer__text">
-                    An independent House, strategically located between Rembrand
-                    Square and National Opera, but where the bustle of the city
-                    comes to rest in this alley flowery and colorful.
-                  </p>
+                  <p className="offer__text">{description}</p>
                 </div>
               </div>
-              <section className="offer__reviews reviews">
-                <h2 className="reviews__title">
-                  Reviews &middot; <span className="reviews__amount">1</span>
-                </h2>
-                <ul className="reviews__list">
-                  <li className="reviews__item">
-                    <div className="reviews__user user">
-                      <div className="reviews__avatar-wrapper user__avatar-wrapper">
-                        <img
-                          className="reviews__avatar user__avatar"
-                          src="img/avatar-max.jpg"
-                          width="54"
-                          height="54"
-                          alt="Reviews avatar"
-                        />
-                      </div>
-                      <span className="reviews__user-name">Max</span>
-                    </div>
-                    <div className="reviews__info">
-                      <div className="reviews__rating rating">
-                        <div className="reviews__stars rating__stars">
-                          <span style={{ width: '80%' }}></span>
-                          <span className="visually-hidden">Rating</span>
-                        </div>
-                      </div>
-                      <p className="reviews__text">
-                        A quiet cozy and picturesque that hides behind a a river
-                        by the unique lightness of Amsterdam. The building is
-                        green and from 18th century.
-                      </p>
-                      <time className="reviews__time" dateTime="2019-04-24">
-                        April 2019
-                      </time>
-                    </div>
-                  </li>
-                </ul>
-                {authorizationStatus === AuthorizationStatus.Auth ? (
-                  <form className="reviews__form form" action="#" method="post">
-                    <label
-                      className="reviews__label form__label"
-                      htmlFor="review"
-                    >
-                      Your review
-                    </label>
-                    <div className="reviews__rating-form form__rating">
-                      <input
-                        className="form__rating-input visually-hidden"
-                        name="rating"
-                        value="5"
-                        id="5-stars"
-                        type="radio"
-                      />
-                      <label
-                        htmlFor="5-stars"
-                        className="reviews__rating-label form__rating-label"
-                        title="perfect"
-                      >
-                        <svg
-                          className="form__star-image"
-                          width="37"
-                          height="33"
-                        >
-                          <use xlinkHref="#icon-star"></use>
-                        </svg>
-                      </label>
-
-                      <input
-                        className="form__rating-input visually-hidden"
-                        name="rating"
-                        value="4"
-                        id="4-stars"
-                        type="radio"
-                      />
-                      <label
-                        htmlFor="4-stars"
-                        className="reviews__rating-label form__rating-label"
-                        title="good"
-                      >
-                        <svg
-                          className="form__star-image"
-                          width="37"
-                          height="33"
-                        >
-                          <use xlinkHref="#icon-star"></use>
-                        </svg>
-                      </label>
-
-                      <input
-                        className="form__rating-input visually-hidden"
-                        name="rating"
-                        value="3"
-                        id="3-stars"
-                        type="radio"
-                      />
-                      <label
-                        htmlFor="3-stars"
-                        className="reviews__rating-label form__rating-label"
-                        title="not bad"
-                      >
-                        <svg
-                          className="form__star-image"
-                          width="37"
-                          height="33"
-                        >
-                          <use xlinkHref="#icon-star"></use>
-                        </svg>
-                      </label>
-
-                      <input
-                        className="form__rating-input visually-hidden"
-                        name="rating"
-                        value="2"
-                        id="2-stars"
-                        type="radio"
-                      />
-                      <label
-                        htmlFor="2-stars"
-                        className="reviews__rating-label form__rating-label"
-                        title="badly"
-                      >
-                        <svg
-                          className="form__star-image"
-                          width="37"
-                          height="33"
-                        >
-                          <use xlinkHref="#icon-star"></use>
-                        </svg>
-                      </label>
-
-                      <input
-                        className="form__rating-input visually-hidden"
-                        name="rating"
-                        value="1"
-                        id="1-star"
-                        type="radio"
-                      />
-                      <label
-                        htmlFor="1-star"
-                        className="reviews__rating-label form__rating-label"
-                        title="terribly"
-                      >
-                        <svg
-                          className="form__star-image"
-                          width="37"
-                          height="33"
-                        >
-                          <use xlinkHref="#icon-star"></use>
-                        </svg>
-                      </label>
-                    </div>
-                    <textarea
-                      className="reviews__textarea form__textarea"
-                      id="review"
-                      name="review"
-                      placeholder="Tell how was your stay, what you like and what can be improved"
-                    >
-                    </textarea>
-                    <div className="reviews__button-wrapper">
-                      <p className="reviews__help">
-                        To submit review please make sure to set{' '}
-                        <span className="reviews__star">rating</span> and
-                        describe your stay with at least{' '}
-                        <b className="reviews__text-amount">50 characters</b>.
-                      </p>
-                      <button
-                        className="reviews__submit form__submit button"
-                        type="submit"
-                        disabled
-                      >
-                        Submit
-                      </button>
-                    </div>
-                  </form>
-                ) : null}
-              </section>
+              {<Reviews />}
             </div>
           </div>
-          <section className="offer__map map"></section>
+
+          {<Map />}
         </section>
         <div className="container">
           <section className="near-places places">
